+++ usaspending_api/awards/v2/views/idvs/funding_treemap.py
@@ -49,10 +49,11 @@ FUNDING_TREEMAP_SQL = SQL("""
             taa.treasury_account_identifier = faba.treasury_account_id
         left outer join federal_account fa on
             taa.agency_id || '-' || taa.main_account_code =  fa.federal_account_code
+        {additional_data_sql}
         {group_by}
         {order_by}
-        {limit}
          {offset};""")
+
 
 
 class IDVFundingBaseViewSet(APIDocumentationView):
@@ -64,7 +65,7 @@ class IDVFundingBaseViewSet(APIDocumentationView):
 
     @staticmethod
     def _business_logic(request_data: dict, columns: object, group_by: object,
-                        limit: object, order_by: object, offset: object) -> list:
+                        order_by: object, offset: object, sql: object ) -> list:
         # By this point, our award_id has been validated and cleaned up by
         # TinyShield.  We will either have an internal award id that is an
         # integer or a generated award id that is a string.
@@ -77,8 +78,8 @@ class IDVFundingBaseViewSet(APIDocumentationView):
             award_id=Literal(award_id),
             group_by=group_by,
             order_by=order_by,
-            limit=limit,
             offset=offset,
+            additional_data_sql=sql
         )
         return execute_sql_to_ordered_dictionary(sql)
 
@@ -92,14 +93,14 @@ class IDVFundingRollupViewSet(IDVFundingBaseViewSet):
 
     @cache_response()
     def post(self, request: Request) -> Response:
-        limit = SQL("")
         order_by = SQL("")
         offset = SQL("")
+        sql = SQL("")
         columns = SQL("""coalesce(sum(nullif(faba.transaction_obligated_amount, 'NaN')), 0.0) total_transaction_obligated_amount,
                      count(distinct ca.awarding_agency_id) awarding_agency_count,
                      count(distinct taa.agency_id || '-' || taa.main_account_code) federal_account_count""")
         group_by = SQL("")
-        results = self._business_logic(request.data, columns, group_by, limit, order_by, offset)
+        results = self._business_logic(request.data, columns, group_by, order_by, offset, sql)
         return Response(results[0])
 
 
@@ -112,19 +113,24 @@ class IDVFundingTreemapViewSet(IDVFundingBaseViewSet):
 
     @cache_response()
     def post(self, request: Request) -> Response:
-        limit = SQL("limit {}". format(request.data['limit'] + 1))
+        sql = SQL("")
         offset = SQL("offset {}".format((request.data['page'] - 1) * request.data['limit']))
         order_by = SQL("order by {} {}".format(SORTABLE_COLUMNS[request.data['sort']], request.data['order']))
-        group_by = SQL("group by federal_account, fa.account_title ")
+        group_by = SQL("group by federal_account, fa.account_title, cte.award_id, ca.funding_agency_id ")
         columns = SQL("""sum(nullif(faba.transaction_obligated_amount, 'NaN')) total_transaction_obligated_amount,
                       taa.agency_id || '-' || taa.main_account_code federal_account,
-                      fa.account_title""")
+                      sum(sum(nullif(faba.transaction_obligated_amount, 'NaN'))) over (partition by cte.award_id) as total,
+                      fa.account_title,
+                      ca.funding_agency_id""")
 
-        results = self._business_logic(request.data, columns, group_by, limit, order_by, offset)
+        results = self._business_logic(request.data, columns, group_by, order_by, offset, sql)
         paginated_results, page_metadata = get_pagination(results, request.data['limit'], request.data['page'])
-
+        total = paginated_results[0]['total'] if len(paginated_results) > 0 else 0
+        for item in paginated_results:
+            del item['total']
         response = OrderedDict((
             ('results', paginated_results),
+            ('total', total),
             ('page_metadata', page_metadata)
         ))
 
@@ -133,12 +139,11 @@ class IDVFundingTreemapViewSet(IDVFundingBaseViewSet):
 
 # STUB FOR DEV-2238
 
-@validate_post_request(TREEMAP_MODELS +
-                       [{'key': 'account',
-                         'name': 'account',
-                         'optional': False,
-                         'type': 'text',
-                         'text_type': 'search'}])
+ACCOUNT_MODEL = [{'key': 'account','name': 'account','optional': False,'type': 'text',
+                  'text_type': 'search'}]
+
+
+@validate_post_request(TREEMAP_MODELS + ACCOUNT_MODEL)                     
 class IDVFundingTreemapDrillDownViewSet(IDVFundingBaseViewSet):
 
     @cache_response()
@@ -150,7 +155,7 @@ class IDVFundingTreemapDrillDownViewSet(IDVFundingBaseViewSet):
         columns = SQL("""sum(nullif(faba.transaction_obligated_amount, 'NaN')) total_transaction_obligated_amount,
                       taa.agency_id || '-' || taa.main_account_code federal_account,""")
 
-        results = self._business_logic(request.data, columns, group_by, limit, order_by, offset)
+        results = self._business_logic(request.data, columns, group_by, order_by, offset)
         paginated_results, page_metadata = get_pagination(results, request.data['limit'], request.data['page'])
 
         response = OrderedDict((
