+++ usaspending_api/settings.py
@@ -223,7 +223,7 @@ USE_TZ = True
 STATIC_URL = '/static/'
 STATIC_ROOT = os.path.join(BASE_DIR, 'usaspending_api/static/')
 STATICFILES_DIRS = (os.path.join(BASE_DIR, 'usaspending_api/static_doc_files'),)
-
+'''
 LOGGING = {
     'version': 1,
     'disable_existing_loggers': False,
@@ -275,7 +275,23 @@ LOGGING = {
         }
     },
 }
-
+'''
+LOGGING = {
+    'version': 1,
+    'handlers': {
+        'console':{
+            'level':'DEBUG',
+            'class':'logging.StreamHandler',
+        },
+    },
+    'loggers': {
+        'django.request': {
+            'handlers':['console'],
+            'propagate': True,
+            'level':'DEBUG',
+        }
+    },
+}
 # If caches added or renamed, edit clear_caches in usaspending_api/etl/helpers.py
 CACHES = {
     'default': {
